Fix undo/redo issues when define[Above,Below] is active
Class ExtrusionRulesProcessor was a bit refactored as well.

TODO. There is still a need to inform the GUI about the status of
define[Above,Below] when undos and redos are issued. The problem is that
the GUI may think that a (for ex.) defineBelow is active even though the
surface that was supposed be the upper boundary for the sketch area may
have been already destroyed by an undo.