Fix case when including files
This commit fix a few misspelled file names in which the case when the
file was included differed from what was written in the file name.
Also, fix a function declaration in the ObjectOpenGLWeapper which
declared an invalid pass by l-reference when a pass by value (or a
universal reference) was required.