Allow coarsening whole model for flow diagnostics
This commit uncomments code that previously allowed to resample the
whole model in a coarser resolution than the one being used for
modelling.  This code was commented out prior to the January 2018
sponsors' meeting because it was broken by truncate, and not revisited
later because HW preferred to try and coarsen the model's individual
surfaces with the triangle library.