Fix bugs in RulesProcessor
Notice that there were a weird issue when trying to read JSON files in
the RulesProcessor::loadFile() method that may happen again.

The stratmod library is ready to cacth any interruption thrown from the
cereal library when errors happen during loading a file, but (for
unknown reasons) such interruptions were killing RRM when debugging in
QtCreator.  It this is the case, comment out the code to load binary
files and just leave the code to load JSON files in RulesProcessor.