Fix bugs in RulesProcessor
There was a logical error when getting sup and inf from bounding curves
in which all curves were supposed to be used, but only those curves for
which at least one segment was valid were actually used.  This led to
discontinuities in the bounding box as viewed from its intersection with
cross-sections.