Update the stratmod library to fix normals' computation
When creating a path guided surface by first sketching the path and then
sketching the cross-section curve, the normals of the created surface
were being incorrectly computed.  The normals of the test surface,
however, were being correctly computed (even though both surfaces should
have been the same).

Fixing a numerical instability in the way derivatives were computed in
the interpolant kernel fixed the visualization, but more tests are
needed to understand what is the difference between the test and the
created surface.