Quickfix (workaround) for ra-rb bug
This commit adds a woraround to a bug in ra-rb that would prevent a
surface to be included in the model when pa-pb is active in a few rare
cases.  The issue stems from how UEnvelopes and LEnvolepes are being
computed in the code, and the source issue has not been addressed as
that piece of code is central to the whole library.