Fix bug in flow diagnostics interface.
The FD interface would not reset the model coordinate system to its
original state if the FD window is closed and reopened again, which
would not trigger the inversion of the y axis when the FD window is
reopened, leading to wrong results.

This commit adds code to force the interface to revert the model to the
OpenGL coordinate system before all calls to the FD interface
createWindow() method.